Further refactor into eflang.core package
