+++ ef-core/src/main/java/eflang/core/Main.java
@@ -1,4 +1,4 @@
-package eflang;
+package eflang.core;
 
 public class Main {
     public static void main(String[] args) {
