+++ ef-core/src/main/java/eflang/core/NullPerformer.java
@@ -1,4 +1,4 @@
-package eflang;
+package eflang.core;
 
 import org.jfugue.MusicStringParser;
