+++ ef-core/src/main/java/eflang/core/SystemIoManager.java
@@ -1,4 +1,4 @@
-package eflang;
+package eflang.core;
 
 import java.util.Scanner;
